Add modular structure and GitHub workflow
- Refactored project structure to include separate files for models and database
- Added a GitHub workflow for building and running the FastAPI application
- Updated main.py to import models and database from new locations
