+++ requirements.txt
@@ -1,0 +1,3 @@
+fastapi==0.70.0
+uvicorn==0.15.0
+fastapi-pagination==0.6.1
