+++ utils/database.py
@@ -1,0 +1,59 @@
+# Hard Coded Database
+raw_trade = [
+    {
+        "assetClass": "Bond",
+        "counterparty": "Hastiin",
+        "instrumentId": "GHJD",
+        "instrumentName": "Debentures",
+        "tradeDateTime": "2024-07-04 00:02:35.520788",
+        "tradeDetails": {
+            "buySellIndicator": "BUY",
+            "price": 9.36,
+            "quantity": 42
+        },
+        "tradeId": "8c8f9d65-ef45-43cc-9253-badac16bdba9",
+        "trader": "Akash"
+    },
+    {
+        "assetClass": "Bond",
+        "counterparty": "Masamba",
+        "instrumentId": "TSWQ",
+        "instrumentName": "Debentures",
+        "tradeDateTime": "2021-02-03 00:03:35.520788",
+        "tradeDetails": {
+            "buySellIndicator": "SELL",
+            "price": 1.36,
+            "quantity": 32
+        },
+        "tradeId": "f0aed2de-d4d2-4213-bed1-60bd6fb1d2b0",
+        "trader": "Otar"
+    },
+    {
+        "assetClass": "Equity",
+        "counterparty": "Fyokla",
+        "instrumentId": "TSEW",
+        "instrumentName": "Preference Shares",
+        "tradeDateTime": "2022-12-03 00:03:35.520788",
+        "tradeDetails": {
+            "buySellIndicator": "BUY",
+            "price": 2.36,
+            "quantity": 12
+        },
+        "tradeId": "8ae396c4-94e3-400b-b44a-ebfc06ef131c",
+        "trader": "Koba"
+    },
+    {
+        "assetClass": "Equity",
+        "counterparty": "Isokrates",
+        "instrumentId": "DAWD",
+        "instrumentName": "Mutual Funds",
+        "tradeDateTime": "2023-10-03 00:06:35.520788",
+        "tradeDetails": {
+            "buySellIndicator": "SELL",
+            "price": 5.36,
+            "quantity": 22
+        },
+        "tradeId": "a3fff395-d13f-4faf-80d6-71b291d43cf0",
+        "trader": "Jaycee"
+    }
+]
